+++ example/lib/ui/home_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:sleek_circular_slider/sleek_circular_slider.dart';
 import 'clock_page.dart';
 import 'random_value_page.dart';
 import 'package:example/utils.dart';
 import 'dart:math' as math;
-import 'package:sleek_circular_slider/sleek_circular_slider.dart';
 import 'example_page.dart';
 
 /// Example 01
