+++ lib/src/painters/circular_arc_painter.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:sleek_circular_slider/sleek_circular_slider.dart';
+import 'package:sleek_circular_slider/src/parameters/circular_slider_values.dart';
 import 'package:sleek_circular_slider/src/utilities/unit_conversions.dart';
 
 abstract class ShapePainter {
