+++ lib/src/parameters/circular_slider_callbacks.dart
@@ -1,0 +1,13 @@
+typedef void OnChange(double value);
+
+class SliderCallbacks {
+  final OnChange onChange;
+  final OnChange onChangeStart;
+  final OnChange onChangeEnd;
+
+  const SliderCallbacks({
+    this.onChange,
+    this.onChangeStart,
+    this.onChangeEnd,
+  });
+}
