+++ lib/src/parameters/circular_slider_painters.dart
@@ -1,21 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:sleek_circular_slider/sleek_circular_slider.dart';
 import 'package:sleek_circular_slider/src/painters/background_painter.dart';
+import 'package:sleek_circular_slider/src/painters/current_value_painter.dart';
+import 'package:sleek_circular_slider/src/painters/progress_bar_painter.dart';
 import 'package:sleek_circular_slider/src/painters/shadow_painter.dart';
+import 'package:sleek_circular_slider/src/parameters/circular_slider_values.dart';
+import 'package:sleek_circular_slider/src/parameters/circular_slider_settings.dart';
 
-typedef void OnChange(double value);
-typedef Widget InnerWidget(double percentage);
 typedef CustomPainter CustomSliderPainter(
   CircularSliderSettings settings,
   CircularSliderValues values,
 );
 
-class SliderPainters {
+class CircularSliderPainters {
   CustomSliderPainter backgroundPainter;
   CustomSliderPainter shadowPainter;
   CustomSliderPainter progressBarPainter;
   CustomSliderPainter currentValuePainter;
-  SliderPainters({
+
+  CircularSliderPainters({
     CustomSliderPainter backgroundPainter,
     CustomSliderPainter shadowPainter,
     CustomSliderPainter progressBarPainter,
@@ -28,16 +30,4 @@ class SliderPainters {
             ((settings, values) => ProgressBarPainter(settings, values)),
         this.currentValuePainter = currentValuePainter ??
             ((settings, values) => CurrentValuePainter(settings, values));
-}
-
-class SliderCallbacks {
-  final OnChange onChange;
-  final OnChange onChangeStart;
-  final OnChange onChangeEnd;
-
-  const SliderCallbacks({
-    this.onChange,
-    this.onChangeStart,
-    this.onChangeEnd,
-  });
 }
