+++ lib/src/parameters/circular_slider_values.dart
