+++ lib/src/settings/circular_slider_colors.dart
