+++ lib/src/settings/circular_slider_features.dart
