+++ lib/src/settings/circular_slider_shadow.dart
