+++ lib/src/settings/circular_slider_text.dart
